+++ assets/swipe/2.0/js/swipe.js
@@ -1,10 +1,12 @@
-/*
+/**
  * Swipe 2.0
  *
  * Brad Birdsall
  * Copyright 2013, MIT License
  *
-*/
+ * Dot menu enhancements
+ * Copyright (c) 2013-2014 Leo Feyer
+ */
 
 function Swipe(container, options) {
 
@@ -41,6 +43,15 @@ function Swipe(container, options) {
 
     // cache slides
     slides = element.children;
+
+    // filter comments (IE8)
+    for (var i=0; i<slides.length; i++) {
+      if (slides[i].nodeType == 8) {
+        element.removeChild(slides[i]);
+        --i;
+      }
+    }
+
     length = slides.length;
 
     // set continuous to false if only one slide
@@ -64,7 +75,7 @@ function Swipe(container, options) {
 
     // stack elements
     var pos = slides.length;
-    while(pos--) {
+    while (pos--) {
 
       var slide = slides[pos];
 
@@ -457,7 +468,7 @@ function Swipe(container, options) {
       }
 
     },
-    end: function(event) {
+    end: function() {
 
       // measure duration
       var duration = +new Date - start.time;
@@ -557,7 +568,7 @@ function Swipe(container, options) {
 
     }
 
-  }
+  };
 
   // trigger setup
   setup();