+++ system/modules/core/dca/tl_layout.php
@@ -416,7 +416,7 @@ $GLOBALS['TL_DCA']['tl_layout'] = array
 			'options_callback'        => array('tl_layout', 'getAnalyticsTemplates'),
 			'reference'               => &$GLOBALS['TL_LANG']['tl_layout'],
 			'eval'                    => array('multiple'=>true),
-			'sql'                     => "varchar(64) NOT NULL default ''"
+			'sql'                     => "text NULL"
 		),
 		'script' => array
 		(