+++ system/modules/core/modules/ModuleArticlenav.php
@@ -78,7 +78,7 @@ class ModuleArticlenav extends \Module
 			}
 
 			$strAlias = ($this->objArticles->alias != '' && !$GLOBALS['TL_CONFIG']['disableAlias']) ? $this->objArticles->alias : $this->objArticles->id;
-			$this->redirect($this->addToUrl('articles=' . $strAlias));
+			$this->redirect($this->generateFrontendUrl($objPage->row(), '/articles/' . $strAlias));
 		}
 
 		return parent::generate();
@@ -90,6 +90,8 @@ class ModuleArticlenav extends \Module
 	 */
 	protected function compile()
 	{
+		global $objPage;
+
 		$intActive = null;
 		$articles = array();
 		$intCount = 1;
@@ -104,7 +106,7 @@ class ModuleArticlenav extends \Module
 				$articles[] = array
 				(
 					'isActive' => true,
-					'href' => $this->addToUrl('articles=' . $strAlias),
+					'href' => $this->generateFrontendUrl($objPage->row(), '/articles/' . $strAlias),
 					'title' => specialchars($this->objArticles->title, true),
 					'link' => $intCount
 				);
@@ -118,7 +120,7 @@ class ModuleArticlenav extends \Module
 				$articles[] = array
 				(
 					'isActive' => false,
-					'href' => $this->addToUrl('articles=' . $strAlias),
+					'href' => $this->generateFrontendUrl($objPage->row(), '/articles/' . $strAlias),
 					'title' => specialchars($this->objArticles->title, true),
 					'link' => $intCount
 				);